+++ src/lib/helpers/__tests__/SwaggerUtils.ts
@@ -1,4 +1,5 @@
 import SwaggerUtils from '../SwaggerUtils';
+import { NodegenRc } from '@/interfaces';
 
 const params = [
   {
@@ -86,36 +87,36 @@ const params = [
 ];
 
 test('Returns joi with 2 required params', () => {
-  expect(SwaggerUtils.createJoiValidation('post', {parameters: [params[0]]})).toBe(
-    `'body': Joi.object({'password':Joi.string().trim(true).required(),'newPassword':Joi.string().trim(true).required(),'newPasswordConfirm':Joi.string().trim(true).allow('').allow(null),}),`
+  expect(SwaggerUtils.createJoiValidation('post', {parameters: [params[0]]}, {} as NodegenRc)).toBe(
+    `'body': Joi.object({'password':Joi.string().required(),'newPassword':Joi.string().required(),'newPasswordConfirm':Joi.string().allow('').allow(null),}),`
   );
 });
 
 test('openapi3 query request param', () => {
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[1]]})).toBe(`'query': Joi.object({'limit':Joi.number().integer(),}),`);
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[1]]}, {} as NodegenRc)).toBe(`'query': Joi.object({'limit':Joi.number().integer(),}),`);
 });
 
 test('openapi2 enums', () => {
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[3]]})).toBe(
-    `'query': Joi.object({'sort':Joi.string().trim(true).allow('').valid('asc', 'desc'),}),`
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[3]]}, {} as NodegenRc)).toBe(
+    `'query': Joi.object({'sort':Joi.string().allow('').valid('asc', 'desc'),}),`
   );
 });
 
 test('openapi3 enums', () => {
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[2]]})).toBe(
-    `'query': Joi.object({'sort':Joi.string().trim(true).allow('').valid('asc', 'desc'),}),`
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[2]]}, {} as NodegenRc)).toBe(
+    `'query': Joi.object({'sort':Joi.string().allow('').valid('asc', 'desc'),}),`
   );
 });
 
 test('openapi3 query request array param: allow single value', () => {
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[4]]})).toBe(
-    `'query': Joi.object({'select':Joi.array().items(Joi.string().trim(true).allow('')).single(),}),`
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[4]]}, {} as NodegenRc)).toBe(
+    `'query': Joi.object({'select':Joi.array().items(Joi.string().allow('')).single(),}),`
   );
 });
 
 test('openapi3 request body: allow empty array', () => {
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[5]]})).toBe(
-    `'body': Joi.object({'selected':Joi.array().items(Joi.string().trim(true).allow('').allow(null)).required(),}),`
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[5]]}, {} as NodegenRc)).toBe(
+    `'body': Joi.object({'selected':Joi.array().items(Joi.string().allow('').allow(null)).required(),}),`
   );
 });
 
@@ -152,15 +153,15 @@ test('add unknown true for headers', () => {
           },
         },
       ],
-    })
+    }, {} as NodegenRc)
   ).toBe(
-    `'headers': Joi.object({'sort':Joi.string().trim(true).valid('asc', 'desc').required(),'access':Joi.string().trim(true).regex(/^Bearer .+$/).required(),}).unknown(true),'query': Joi.object({'limit':Joi.number().integer(),}),`
+    `'headers': Joi.object({'sort':Joi.string().valid('asc', 'desc').required(),'access':Joi.string().regex(/^Bearer .+$/).required(),}).unknown(true),'query': Joi.object({'limit':Joi.number().integer(),}),`
   );
 });
 
 test('query item key is properly created inside the query object', () => {
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[6]]})).toBe(
-    `'query': Joi.object({'objectInsideQuery': Joi.object({'prop1':Joi.string().trim(true).allow(''),}),}),`
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [params[6]]}, {} as NodegenRc)).toBe(
+    `'query': Joi.object({'objectInsideQuery': Joi.object({'prop1':Joi.string().allow(''),}),}),`
   );
 });
 
@@ -175,21 +176,21 @@ test('query numeric with min / max zero properly adds validation', () => {
     minimum: 0,
   };
 
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe('\'query\': Joi.object({\'limit\':Joi.number().min(0),}),');
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe('\'query\': Joi.object({\'limit\':Joi.number().min(0),}),');
 
   param.type = 'integer';
   param.schema.type = 'integer';
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe('\'query\': Joi.object({\'limit\':Joi.number().integer().min(0),}),');
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe('\'query\': Joi.object({\'limit\':Joi.number().integer().min(0),}),');
 
   param.type = 'number';
   param.schema.type = 'number';
   param.schema.maximum = 0;
   param.maximum = 0;
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe('\'query\': Joi.object({\'limit\':Joi.number().min(0).max(0),}),');
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe('\'query\': Joi.object({\'limit\':Joi.number().min(0).max(0),}),');
 
   param.type = 'integer';
   param.schema.type = 'integer';
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe(
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe(
     '\'query\': Joi.object({\'limit\':Joi.number().integer().min(0).max(0),}),'
   );
 });
@@ -205,12 +206,12 @@ test('query string with min / max zero properly adds validation', () => {
     minimum: 0,
   };
 
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe('\'query\': Joi.object({\'limit\':Joi.string().trim(true).allow(\'\').min(0),}),');
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe('\'query\': Joi.object({\'limit\':Joi.string().allow(\'\').min(0),}),');
 
   param.schema.maximum = 0;
   param.maximum = 0;
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe(
-    '\'query\': Joi.object({\'limit\':Joi.string().trim(true).allow(\'\').min(0).max(0),}),'
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe(
+    '\'query\': Joi.object({\'limit\':Joi.string().allow(\'\').min(0).max(0),}),'
   );
 });
 
@@ -253,8 +254,8 @@ test('respects x-nullable and nullable for required body parameters', () => {
     },
   };
 
-  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]})).toBe(
-    `'body': Joi.object({'email':Joi.string().trim(true).required(),'phoneNumber':Joi.string().trim(true).allow('').allow(null).required(),'comment':Joi.string().trim(true).allow('').allow(null).required(),'info':Joi.string().trim(true).min(10).allow(null).required(),'code':Joi.number().allow(null).required(),}),`
+  expect(SwaggerUtils.createJoiValidation('get', {parameters: [param]}, {} as NodegenRc)).toBe(
+    `'body': Joi.object({'email':Joi.string().required(),'phoneNumber':Joi.string().allow('').allow(null).required(),'comment':Joi.string().allow('').allow(null).required(),'info':Joi.string().min(10).allow(null).required(),'code':Joi.number().allow(null).required(),}),`
   );
 });
 
@@ -301,9 +302,9 @@ test('post object with nested objects that are required should be required in th
     ]
   };
 
-  const joiString = SwaggerUtils.createJoiValidation('post', pathObj);
+  const joiString = SwaggerUtils.createJoiValidation('post', pathObj, {} as NodegenRc);
 
-  const expectedouput = `'body': Joi.object({'team':Joi.object({'name':Joi.string().trim(true).min(1).required(),'isPremium':Joi.boolean().required(),}).allow(null),'invitations':Joi.array().items(Joi.object({'id':Joi.number().allow(null),'name':Joi.string().trim(true).allow('').allow(null),'email':Joi.string().trim(true).allow('').allow(null),'roleId':Joi.number().required(),'invitationLink':Joi.string().trim(true).allow('').allow(null),'action':Joi.string().trim(true).allow('').valid('create', 'update', 'delete').allow(null),}).allow(null)),'invitationsLanguage':Joi.string().trim(true).allow('').allow(null),}),`;
+  const expectedouput = `'body': Joi.object({'team':Joi.object({'name':Joi.string().min(1).required(),'isPremium':Joi.boolean().required(),}).allow(null),'invitations':Joi.array().items(Joi.object({'id':Joi.number().allow(null),'name':Joi.string().allow('').allow(null),'email':Joi.string().allow('').allow(null),'roleId':Joi.number().required(),'invitationLink':Joi.string().allow('').allow(null),'action':Joi.string().allow('').valid('create', 'update', 'delete').allow(null),}).allow(null)),'invitationsLanguage':Joi.string().allow('').allow(null),}),`;
 
   expect(joiString).toBe(expectedouput);
 });
@@ -335,13 +336,60 @@ it('When the string attribute contains x-dont-trim then trim(true) should not be
     ]
   };
 
-  const joiString = SwaggerUtils.createJoiValidation('post', pathObj);
+  const joiString = SwaggerUtils.createJoiValidation('post', pathObj, {} as NodegenRc);
 
   const expectedouput = `'body': Joi.object({'team':Joi.object({'name':Joi.string().min(1).required(),}).allow(null),}),`;
 
   expect(joiString).toBe(expectedouput);
 });
 
+it('Default trim from nodegenrc always opt-out and off', () => {
+  let pathObj: any = {
+    'parameters': [
+      {
+        in: 'body',
+        name: 'teamsPostPost',
+        required: true,
+        schema: {
+          type: 'object',
+          required: ['name'],
+          properties: {
+            name: {
+              type: 'string',
+            },
+          },
+        },
+      },
+    ]
+  };
+
+  expect(
+    SwaggerUtils.createJoiValidation('post', pathObj, {joi: {strings: {autoTrim: 'always'}}} as NodegenRc)
+  ).toBe(
+    `'body': Joi.object({'name':Joi.string().trim(true).required(),}),`
+  );
+
+  expect(
+    SwaggerUtils.createJoiValidation('post', pathObj, {joi: {strings: {autoTrim: 'off'}}} as NodegenRc)
+  ).toBe(
+    `'body': Joi.object({'name':Joi.string().required(),}),`
+  );
+
+  expect(
+    SwaggerUtils.createJoiValidation('post', pathObj, {joi: {strings: {autoTrim: 'opt-out'}}} as NodegenRc)
+  ).toBe(
+    `'body': Joi.object({'name':Joi.string().trim(true).required(),}),`
+  );
+
+  // add the don't trim to test the opt-out
+  pathObj.parameters[0].schema.properties.name['x-dont-trim'] = true;
+
+  expect(
+    SwaggerUtils.createJoiValidation('post', pathObj, {joi: {strings: {autoTrim: 'opt-out'}}} as NodegenRc)
+  ).toBe(
+    `'body': Joi.object({'name':Joi.string().required(),}),`
+  );
+});
 
 it('Adding the Joi methods to the string', () => {
   const pathObj: any = {
@@ -360,8 +408,10 @@ it('Adding the Joi methods to the string', () => {
                 name: {
                   type: 'string',
                   minLength: 1,
-                  'x-joi-string-lowercase': true,
-                  'x-joi-string-email': true,
+                  'x-joi-lowercase': null,
+                  'x-joi-email': {
+                    allowFullyQualified: true
+                  },
                 },
               },
             }
@@ -371,9 +421,9 @@ it('Adding the Joi methods to the string', () => {
     ]
   };
 
-  const joiString = SwaggerUtils.createJoiValidation('post', pathObj);
+  const joiString = SwaggerUtils.createJoiValidation('post', pathObj, {} as NodegenRc);
 
-  const expectedouput = `'body': Joi.object({'team':Joi.object({'name':Joi.string().trim(true).lowercase().email().min(1).required(),}).allow(null),}),`;
+  const expectedouput = `'body': Joi.object({'team':Joi.object({'name':Joi.string().lowercase().email({"allowFullyQualified":true}).min(1).required(),}).allow(null),}),`;
 
   expect(joiString).toBe(expectedouput);
 });